fix: correct remaining int paths, document SQLite migration, add >int.MaxValue parser tests (#183)
- Rename ParseInt to ParseLong in Stockpiles.razor (returns long, not int)
- Use 0L literal in ParseLong/ParseQuantity defaults
- Fix overflow bug in SumAboveIntMaxValue test ((int)bigStack*2 → bigStack*2)
- Add explanatory comment in empty migration: SQLite INTEGER is already 64-bit;
  EF snapshot captures the model change; no ALTER TABLE needed
- Add 4 bulk parser regression tests: >int.MaxValue, near long.MaxValue,
  overflow rejection, int.MaxValue+1 acceptance

+++ tests/WALLEve.Tests/StockpileCalculationTests.cs
@@ -327,7 +327,7 @@ public class StockpileCalculationTests
         long bigStack = 1_500_000_001L;
         long expectedTotal = 3_000_000_002L;
 
-        var targets = new[] { Target(typeId: 34, quantity: (int)bigStack * 2) };
+        var targets = new[] { Target(typeId: 34, quantity: bigStack * 2) };
         var assets = new List<StockpileCalculator.AssetLine>
         {
             new(ItemId: 100, TypeId: 34, LocationId: 60003760, Quantity: bigStack),
